+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Full documentation for ROCm Systems Profiler is available at [https://rocm.docs.
 
 ### Added
 
+- Support for UCX (Unified Communication X) API tracing.
 - Documentation for `--trace-legacy` / `-L` CLI flag for direct tracing mode.
 - Added dependency to `spdlog` library.
 - Added environment variable `ROCPROFSYS_LOG_LEVEL` which control level of logging.